PET-1108: fix: prevent disk size overflow in rcNeeded calculation

+++ web/src/routes/migration.go
@@ -140,7 +140,7 @@ func (a *MigrationAdmin) Create(ctx context.Context, name string, instances []*m
 				err = nil
 				continue
 			}
-			rcNeeded := fmt.Sprintf("cpu=%d memory=%d disk=%d network=%d", instance.Cpu, instance.Memory*1024, instance.Disk*1024*1024, 0)
+			rcNeeded := fmt.Sprintf("cpu=%d memory=%d disk=%d network=%d", instance.Cpu, instance.Memory*1024, int64(instance.Disk)*1024*1024, 0)
 			control = "select=" + hyperGroup + rcNeeded
 		}
 		err = db.Model(instance).Update("status", model.InstanceStatusMigrating).Error
